TIP-706: move text integration tests
