+++ src/Pim/Bundle/CatalogBundle/tests/integration/ElasticSearch/IndexConfiguration/PimCatalogTextIntegration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\tests\integration\Elasticsearch;
+namespace Pim\Bundle\CatalogBundle\tests\integration\Elasticsearch\IndexConfiguration;
 
 /**
  * This integration tests checks that given an index configuration and some products indexed
@@ -10,7 +10,7 @@ namespace Pim\Bundle\CatalogBundle\tests\integration\Elasticsearch;
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PimCatalogTextIndexConfigurationIntegration extends AbstractIndexConfigurationIntegration
+class PimCatalogTextIndexConfigurationIntegration extends AbstractPimCatalogIntegration
 {
     public function testStartWithOperator()
     {
